correct type

+++ viz_corruption.py
@@ -74,8 +74,8 @@ def main(config, args):
                     print("euclidean difference: ", {-(pre_corrupt.unsqueeze(2) -
                        post_corrupt.unsqueeze(1)).pow(2).sum(dim=-1)})
 
-                    plt.imsave(f'imgs/{idx}_pre_corrupt',pre_corrupt.type(dtype=torch.int).permute(1, 2, 0).cpu().numpy()  )
-                    plt.imsave(f'imgs/{idx}_post_corrupt',post_corrupt.type(dtype=torch.int).permute(1, 2, 0).cpu().numpy()  )
+                    plt.imsave(f'imgs/{idx}_pre_corrupt',pre_corrupt.type(dtype=torch.uint8).permute(1, 2, 0).cpu().numpy()  )
+                    plt.imsave(f'imgs/{idx}_post_corrupt',post_corrupt.type(dtype=torch.uint8).permute(1, 2, 0).cpu().numpy()  )
